Updated minion command

+++ app/src/Command/InstallSlaves.php
@@ -4,7 +4,9 @@ namespace Omgcatz\Command;
 
 use Knp\Command\Command;
 use Omgcatz\Includes\Database;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class InstallSlaves extends Command
@@ -13,7 +15,9 @@ class InstallSlaves extends Command
   {
     $this
       ->setName('setup:slaves')
-      ->setDescription('Install slaves');
+      ->setDescription('Install slaves')
+      ->addArgument('minions', InputArgument::IS_ARRAY, 'The server names/IPs for minions')
+      ->addOption('clear', null, InputOption::VALUE_NONE, 'Clear existing minions from the database');
   }
 
   protected function execute(InputInterface $input, OutputInterface $output)
@@ -21,15 +25,17 @@ class InstallSlaves extends Command
     /** @var Database $db */
     $db = $this->getSilexApplication()['database'];
 
+    $minions = $input->getArgument('minions');
+    $clear = $input->getOption('clear');
 
-    $db->simpleQuery("TRUNCATE TABLE minions");
 
-    // TODO: Minions
+    if ($clear !== null) {
+      $db->simpleQuery("TRUNCATE TABLE minions");
+    }
 
-//    foreach (Config::$minions as $minion) {
-//      $db->simpleQuery("INSERT INTO minions (`minionRoot`,`load`) VALUES ('{$minion}',0)");
-//    }
-
+    foreach ($minions as $minion) {
+      $db->simpleQuery("INSERT INTO minions (`minionRoot`,`load`) VALUES ('{$minion}',0)");
+    }
 
     $output->writeln('Minons Created');
   }
